requirements.txt
fastapi==0.95.2
pydantic==1.10.2
typing-extensions==4.4.0
pymongo==4.3.3
uvicorn==0.30.6
email-validator
python-dotenv